+++ starsky/starsky/Controllers/HealthController.cs
@@ -14,6 +14,7 @@ using starsky.foundation.platform.Extensions;
 using starsky.foundation.platform.VersionHelpers;
 using starsky.foundation.webtelemetry.Helpers;
 using starsky.foundation.webtelemetry.Interfaces;
+using starsky.foundation.webtelemetry.Services;
 using starskycore.ViewModels;
 
 [assembly: InternalsVisibleTo("starskytest")]
