+++ starsky/starskytest/Controllers/HealthControllerTest.cs
@@ -8,6 +8,7 @@ using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using starsky.Controllers;
 using starsky.foundation.webtelemetry.Helpers;
+using starsky.foundation.webtelemetry.Services;
 using starskycore.ViewModels;
 using starskytest.FakeMocks;
