+++ starsky/starskytest/starsky.foundation.webtelemetry/Helpers/ApplicationInsightsJsHelperTest.cs
@@ -23,6 +23,7 @@ using starsky.foundation.database.Data;
 using starsky.foundation.injection;
 using starsky.foundation.platform.Models;
 using starsky.foundation.webtelemetry.Helpers;
+using starsky.foundation.webtelemetry.Services;
 using starskycore.Helpers;
 
 namespace starskytest.Helpers
